requirements.txt
torch==1.7.1 
torchvision==0.8.2 
torch-scatter==2.0.6
torch-sparse==0.6.9 
torch-geometric==1.6.3
PyYAML
rdkit-pypi==2020.9.5.2
tensorboard
